Refactored some code. Rest api is ready to work
