+++ Web App/GB_Webpage/GB_Webpage/Models/ArticleModel.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.EntityFrameworkCore;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 
 namespace GB_Webpage.Models
 {
