+++ Web App/GB_Webpage/GB_Webpage/Properties/launchSettings.json
@@ -15,7 +15,8 @@
       "applicationUrl": "https://localhost:7185;http://localhost:5230",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development"
-      }
+      },
+      "DaysValidToken": 1
     },
     "IIS Express": {
       "commandName": "IISExpress",
